requirements.txt
python-dotenv~=0.19.0
discord_webhook~=0.14.0